Introduced BindingConfigs
The IBinder.Bind method now accepts a BindingConfi instance that
contains information for the binder on how it is expected to
behave during binding. Additionally the Bind(), Bind<>() and BindTo()
extensions enables you to pass this in.

Currently the only piece of configuration that is in the BindingConfig
type is to control whether or not the binder should be allowed to
overwrite properties that doesn't have default(T) as its value.

Additionally there are two static members on it called Default and
AllowOverride. The latter returns an instance of BindingConfig with
Override set to true.

Fixes #822