Return 404 instread of diag help page in release
Updates so that StatisConfiguration.IsRunningDebug is public. The
DiagnosticsHook will not check that flag before returning the
diagnostics help view. If true then it will return NotFound
instead.

Also updated the diagnostics settings page so that it only lists
members, of StaticConfiguration, that has been decorated with a
DescriptionAttribute.

Fixes #625